DAOS-4447 control: organise server mgmt svc files
Organise mgmt_*.go files to make navigation easier.
Add rank filtering to system package.


Conflicts:
	src/control/server/mgmt_svc_test.go
