+++ src/control/system/member_test.go
@@ -216,6 +216,85 @@ func TestMember_AddOrUpdate(t *testing.T) {
 	}
 }
 
+func TestMember_HostRanks(t *testing.T) {
+	addr1, err := net.ResolveTCPAddr("tcp", "127.0.0.1:10001")
+	if err != nil {
+		t.Fatal(err)
+	}
+	members := Members{
+		mockMember(t, 1, MemberStateStarted),
+		mockMember(t, 2, MemberStateStopped),
+		mockMember(t, 3, MemberStateEvicted),
+		NewMember(Rank(4), "", addr1, MemberStateStopped), // second host rank
+	}
+
+	for name, tc := range map[string]struct {
+		members      Members
+		rankList     []Rank
+		expRanks     []Rank
+		expHostRanks map[string][]Rank
+		expMembers   Members
+	}{
+		"no rank list": {
+			members:  members,
+			expRanks: []Rank{1, 2, 3, 4},
+			expHostRanks: map[string][]Rank{
+				"127.0.0.1:10001": {Rank(1), Rank(4)},
+				"127.0.0.2:10001": {Rank(2)},
+				"127.0.0.3:10001": {Rank(3)},
+			},
+			expMembers: members,
+		},
+		"subset rank list": {
+			members:  members,
+			rankList: []Rank{1, 2},
+			expRanks: []Rank{1, 2, 3, 4},
+			expHostRanks: map[string][]Rank{
+				"127.0.0.1:10001": {Rank(1)},
+				"127.0.0.2:10001": {Rank(2)},
+			},
+			expMembers: Members{
+				mockMember(t, 1, MemberStateStarted),
+				mockMember(t, 2, MemberStateStopped),
+			},
+		},
+		"distinct rank list": {
+			members:      members,
+			rankList:     []Rank{0, 5},
+			expRanks:     []Rank{1, 2, 3, 4},
+			expHostRanks: map[string][]Rank{},
+		},
+		"superset rank list": {
+			members:  members,
+			rankList: []Rank{0, 1, 2, 3, 4, 5},
+			expRanks: []Rank{1, 2, 3, 4},
+			expHostRanks: map[string][]Rank{
+				"127.0.0.1:10001": {Rank(1), Rank(4)},
+				"127.0.0.2:10001": {Rank(2)},
+				"127.0.0.3:10001": {Rank(3)},
+			},
+			expMembers: members,
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			log, buf := logging.NewTestLogger(t.Name())
+			defer ShowBufferOnFailure(t, buf)
+
+			ms := NewMembership(log)
+
+			for _, m := range tc.members {
+				if _, err := ms.Add(m); err != nil {
+					t.Fatal(err)
+				}
+			}
+
+			AssertEqual(t, tc.expRanks, ms.Ranks(), "ranks")
+			AssertEqual(t, tc.expHostRanks, ms.HostRanks(tc.rankList...), "host ranks")
+			AssertEqual(t, tc.expMembers, ms.Members(tc.rankList), "members")
+		})
+	}
+}
+
 func TestMember_Convert(t *testing.T) {
 	membersIn := Members{mockMember(t, 1, MemberStateStarted)}
 	membersOut := Members{}
@@ -223,32 +302,6 @@ func TestMember_Convert(t *testing.T) {
 		t.Fatal(err)
 	}
 	AssertEqual(t, membersIn, membersOut, "")
-}
-
-func TestMember_RanksHostsMembers(t *testing.T) {
-	log, buf := logging.NewTestLogger(t.Name())
-	defer ShowBufferOnFailure(t, buf)
-
-	ms := NewMembership(log)
-
-	members := Members{
-		mockMember(t, 1, MemberStateStarted),
-		mockMember(t, 2, MemberStateStopped),
-		mockMember(t, 3, MemberStateEvicted),
-	}
-	for _, m := range members {
-		if _, err := ms.Add(m); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	AssertEqual(t, []Rank{1, 2, 3}, ms.Ranks(), "ranks")
-	AssertEqual(t, map[string][]Rank{
-		"127.0.0.1:10001": {1},
-		"127.0.0.2:10001": {2},
-		"127.0.0.3:10001": {3},
-	}, ms.HostRanks(), "host ranks")
-	AssertEqual(t, members, ms.Members(), "members")
 }
 
 func TestMemberResult_Convert(t *testing.T) {
@@ -302,7 +355,7 @@ func TestMember_UpdateMemberStates(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	for i, m := range ms.Members() {
+	for i, m := range ms.Members([]Rank{}) {
 		AssertEqual(t, expStates[i], m.State(), "")
 	}
 }
